WM-3033 : Importing a Rest service fails with Null Pointer error
Fixed

git-svn-id: https://src.springsource.com/svn/wm_svn/eng3/trunk@32755 bab4cc61-abf4-0310-9601-8b162d8b0fb1

+++ dev/core/tools/src/com/wavemaker/tools/ws/RESTServiceGenerator.java
@@ -100,11 +100,13 @@ public class RESTServiceGenerator extends WebServiceGenerator {
                 String fileName = entry.getKey().substring(0, 1).toUpperCase() + entry.getKey().substring(1)
                         + "result" + StringUtils.JAVA_SRC_EXT;
                 File targetFile = new File(fpath, fileName);
-                String content = FileUtils.readFileToString(targetFile, ServerConstants.DEFAULT_ENCODING);
-                String fromStr = "@XmlRootElement(name = \"" + entry.getValue().getName() + "result\")";
-                String toStr = "@XmlRootElement(name = \"result\")";
-                content = content.replace(fromStr, toStr);
-                FileUtils.writeStringToFile(targetFile, content, ServerConstants.DEFAULT_ENCODING);
+                if (targetFile.exists()) {
+                    String content = FileUtils.readFileToString(targetFile, ServerConstants.DEFAULT_ENCODING);
+                    String fromStr = "@XmlRootElement(name = \"" + entry.getValue().getName() + "result\")";
+                    String toStr = "@XmlRootElement(name = \"result\")";
+                    content = content.replace(fromStr, toStr);
+                    FileUtils.writeStringToFile(targetFile, content, ServerConstants.DEFAULT_ENCODING);
+                }
             }
         } catch (IOException ex) {
             throw new GenerationException(ex);
